+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Updated all depenedncies to latest version
+- Updated all dependencies to latest version
 
 ## [1.51.0] - 2024-06-24