+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a test case to check the status of all Jobs in the WC
+
 ### Changed
 
 - Use actual types from Cert-Manager instead of using `unstructured` for `Certificate` and `ClusterIssuer`
+- Refactored some functions to explicitly take in a context instance instead of using `context.Background()` so that timeouts set on the context can be respected
 
 ## [1.52.0] - 2024-06-25